backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
pytest==7.4.3
pytest-flask==1.3.0
